requirements.txt
asgiref==3.5.2
black==22.12.0
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
Django==4.1.4
django-debug-toolbar==3.8.1
djangorestframework==3.14.0
environ==1.0
filelock==3.8.2
identify==2.5.9
isort==5.11.2
mypy-extensions==0.4.3
nodeenv==1.7.0
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.0
pre-commit==2.20.0
psycopg2-binary==2.9.5
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typing-extensions==4.4.0
virtualenv==20.17.1
